The WindowType was moved to its own module
